+++ src/main/java/me/rexsystems/rexChat/service/PreviewGuiService.java
@@ -1,9 +1,11 @@
 package me.rexsystems.rexChat.service;
 
 import me.rexsystems.rexChat.RexChat;
+import me.rexsystems.rexChat.listener.PreviewGuiListener;
 import me.rexsystems.rexChat.utils.ColorUtils;
 import me.rexsystems.rexChat.utils.MessageUtils;
 import me.rexsystems.rexChat.utils.SchedulerUtils;
+import me.rexsystems.rexChat.utils.ShulkerBoxUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
 import org.bukkit.entity.Player;
@@ -132,6 +134,25 @@ public class PreviewGuiService {
         for (int i = 0; i < 9; i++)
             if (i != 4)
                 inv.setItem(i, filler);
+
+        SchedulerUtils.runForPlayer(plugin, viewer, () -> viewer.openInventory(inv));
+    }
+
+    public void openShulkerPreview(Player viewer, ItemStack shulker) {
+        if (viewer == null || !ShulkerBoxUtils.isShulkerBox(shulker)) {
+            return;
+        }
+
+        String title = ShulkerBoxUtils.getDisplayTitle(shulker);
+        Inventory inv = Bukkit.createInventory(new PreviewGuiListener.PreviewGuiHolder(), 27, title);
+
+        ItemStack[] contents = ShulkerBoxUtils.getContents(shulker);
+        for (int i = 0; i < Math.min(contents.length, 27); i++) {
+            ItemStack item = contents[i];
+            if (item != null && item.getType() != Material.AIR && item.getAmount() > 0) {
+                inv.setItem(i, item.clone());
+            }
+        }
 
         SchedulerUtils.runForPlayer(plugin, viewer, () -> viewer.openInventory(inv));
     }